fix: 회비 입금 엔드포인트 수정
