+++ src/main/java/com/github/kellyihyeon/stanceadmin/application/deposit/DepositService.java
@@ -1,11 +1,21 @@
 package com.github.kellyihyeon.stanceadmin.application.deposit;
 
+import com.github.kellyihyeon.stanceadmin.application.accountbook.dto.MembershipFeeByGuest;
+import com.github.kellyihyeon.stanceadmin.application.accountbook.dto.MembershipFeeForm;
 import com.github.kellyihyeon.stanceadmin.application.deposit.dto.*;
+import com.github.kellyihyeon.stanceadmin.application.member.dto.MemberIdAndName;
+import com.github.kellyihyeon.stanceadmin.domain.accountbook.AccountBook;
+import com.github.kellyihyeon.stanceadmin.domain.accountbook.TransactionType;
+import com.github.kellyihyeon.stanceadmin.domain.deposit.Deposit;
 import com.github.kellyihyeon.stanceadmin.domain.deposit.DepositCategory;
 import com.github.kellyihyeon.stanceadmin.domain.member.MemberType;
+import com.github.kellyihyeon.stanceadmin.domain.member.MembershipFeeType;
+import com.github.kellyihyeon.stanceadmin.infrastructure.repository.accountbook.AccountBookRepository;
 import com.github.kellyihyeon.stanceadmin.infrastructure.repository.deposit.DepositRepository;
 import com.github.kellyihyeon.stanceadmin.infrastructure.repository.member.MemberRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -14,13 +24,76 @@ import java.time.Month;
 import java.time.Year;
 import java.util.*;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class DepositService {
 
     private final MemberRepository memberRepository;
     private final DepositRepository depositRepository;
+    private final AccountBookRepository accountBookRepository;
 
+
+    @Transactional
+    public void registerMembershipFeeByMember(MembershipFeeForm membershipFeeForm) {
+        List<Deposit> deposits = Deposit.toEntityList(membershipFeeForm);
+        saveDeposits(deposits);
+    }
+
+    @Transactional
+    public void registerMembershipFeeByGuest(MembershipFeeByGuest membershipFeeByGuest) {
+        // TODO. 회원으로 등록하기
+        List<MemberIdAndName> members = new LinkedList<>();
+        members.add(new MemberIdAndName(998L, "coffee"));
+
+        List<Deposit> deposits = Deposit.toEntityList(
+                new MembershipFeeForm(
+                        MemberType.GUEST,
+                        MembershipFeeType.ONE_DAY,
+                        membershipFeeByGuest.getAmount(),
+                        membershipFeeByGuest.getDueMonth(),
+                        membershipFeeByGuest.getDepositDate(),
+                        members,
+                        membershipFeeByGuest.getDescription()
+                )
+        );
+
+        saveDeposits(deposits);
+    }
+
+    public void registerExtraFeeByMember(ExtraFee extraFee) {
+        List<Deposit> deposits = Deposit.toEntityList(extraFee);
+        saveDeposits(deposits);
+    }
+
+    private void saveDeposits(List<Deposit> deposits) {
+        for (Deposit deposit : deposits) {
+            depositRepository.save(deposit);
+            updateBalance(deposit);
+        }
+    }
+
+    // TODO. Refactoring 대상
+    private void updateBalance(Deposit deposit) {
+        AccountBook latestAccountBook = accountBookRepository.findTopByOrderByIdDesc().orElseThrow(() -> new IllegalStateException("입출금 데이터가 없습니다."));
+        log.debug("스탠스 가계부 최신 데이터 [{}]의 잔액 [{}]", latestAccountBook.getId(), latestAccountBook.getBalance());
+
+        AccountBook updatedAccountBook = latestAccountBook.updateBalance(
+                deposit.getId(),
+                TransactionType.DEPOSIT,
+                deposit.getDepositDate(),
+                deposit.getDepositor(),
+                deposit.getAmount()
+        );
+
+        accountBookRepository.save(updatedAccountBook);
+    }
+
+    public void registerCashByBank(CashFormByBank cashFormByBank) {
+        Deposit deposit = cashFormByBank.toEntity();
+        depositRepository.save(deposit);
+        updateBalance(deposit);
+    }
 
     public MembershipFeePaidMemberResponse getMembershipFeePaidMemberStatistics(Year year, Month month) {
         Long regularMembers = memberRepository.countByMemberTypeIn(Arrays.asList(MemberType.ACTIVE, MemberType.INACTIVE));
