+++ src/main/java/com/github/kellyihyeon/stanceadmin/presentation/deposit/DepositController.java
@@ -1,6 +1,10 @@
 package com.github.kellyihyeon.stanceadmin.presentation.deposit;
 
+import com.github.kellyihyeon.stanceadmin.application.accountbook.dto.MembershipFeeByGuest;
+import com.github.kellyihyeon.stanceadmin.application.accountbook.dto.MembershipFeeForm;
 import com.github.kellyihyeon.stanceadmin.application.deposit.DepositService;
+import com.github.kellyihyeon.stanceadmin.application.deposit.dto.CashFormByBank;
+import com.github.kellyihyeon.stanceadmin.application.deposit.dto.ExtraFee;
 import com.github.kellyihyeon.stanceadmin.application.deposit.dto.MembershipFeePaidMemberResponse;
 import com.github.kellyihyeon.stanceadmin.application.deposit.dto.TopDepositCategoriesResponse;
 import lombok.RequiredArgsConstructor;
@@ -12,13 +16,41 @@ import java.time.Year;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/deposit")
+@RequestMapping("/deposits")
 public class DepositController {
 
     private final DepositService depositService;
 
+    @PostMapping("/members/membership-fee")
+    void depositMembershipFeeByMember(
+            @RequestBody MembershipFeeForm membershipFeeForm
+    ) {
+        depositService.registerMembershipFeeByMember(membershipFeeForm);
+    }
 
-    @GetMapping("/membership-fee/{year}/{month}")
+    @PostMapping("/guests/membership-fee")
+    void depositMembershipFeeByGuest(
+            @RequestBody MembershipFeeByGuest membershipFeeByGuest
+    ) {
+        depositService.registerMembershipFeeByGuest(membershipFeeByGuest);
+    }
+
+    @PostMapping("/members/extra-fee")
+    void depositExtraFeeByMember(
+            @RequestBody ExtraFee extraFee
+    ) {
+        depositService.registerExtraFeeByMember(extraFee);
+    }
+
+    @PostMapping("/banks")
+    void depositCashByBank(
+            @RequestBody CashFormByBank cashFormByBank
+    ) {
+        depositService.registerCashByBank(cashFormByBank);
+    }
+
+
+    @GetMapping("/membership-fees/{year}/{month}")
     public ResponseEntity<MembershipFeePaidMemberResponse> getMembershipFeePaidMemberStatistics(
             @PathVariable Year year,
             @PathVariable Month month
